+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Fixed ###
 
+  * [retail] Error when clicking optional reagents
   * [retail] Achievement search preview backdrop was broken
   * [retail] Adventure rewards weren't skinned
   * [retail] SSAO graphics option was not skinned