+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿## [9.0.2.0] ##
+﻿## [9.0.2.1] ##
 ### Added ###
 
   * [retail] Campaign Recap
@@ -14,6 +14,12 @@
   * [retail] Error when clicking optional reagents
   * [retail] Achievement search preview backdrop was broken
   * [retail] Adventure rewards weren't skinned
+
+
+
+## [9.0.2.0] ##
+### Fixed ###
+
   * [retail] SSAO graphics option was not skinned
   * [retail] The dungeon ready popup had two backdrops
   * [retail] The world map's close button was out of place when maximized
@@ -291,6 +297,7 @@
   * Error when opening azerite respec UI
 
 [Unreleased]: https://github.com/Haleth/Aurora/compare/master...develop
+[9.0.2.1]: https://github.com/Haleth/Aurora/compare/9.0.2.0...9.0.2.1
 [9.0.2.0]: https://github.com/Haleth/Aurora/compare/9.0.1.7...9.0.2.0
 [9.0.1.7]: https://github.com/Haleth/Aurora/compare/9.0.1.6...9.0.1.7
 [9.0.1.6]: https://github.com/Haleth/Aurora/compare/9.0.1.5...9.0.1.6